pyproject.toml
[project]
name = "voice-test-harness"
version = "0.1.0"
description = "Autonomous voice agent testing framework"
requires-python = ">=3.12"
dependencies = [
    "livekit",
    "livekit-agents>=1.4.0",
    "livekit-plugins-cartesia",
    "livekit-plugins-openai",
    "livekit-plugins-silero",
    "livekit-plugins-deepgram",
    "python-dotenv",
    "pyyaml",
    "jiwer",
    "httpx",
]

[tool.uv]
dev-dependencies = [
    "ruff",
]
